// bot-server/type.ts
export interface AllGroupsResponse {
  success: boolean;
  msg: string;
  groups: Group[];
  status: number;
}

export interface Group {
  id: string;
  group_user_id: string;
  group_name: string;
  has_password: number;
  group_pfp: string;
  participants: Participant[];
  created_at: Date;
  updated_at: Date;
  broadcaster: Broadcaster;
  group_settings: GroupSettings;
}

export interface Broadcaster {
  id: string;
  group_id: string;
  motd_enabled: boolean;
  motd: null | string;
  motd_time: Date | null | string;
  pfpslide_enabled: boolean;
  pfp_slide: null | string;
  pfp_slide_interval: number | null;
  created_at: Date;
  updated_at: Date;
}

export interface GroupSettings {
  id: string;
  group_id: string;
  lock_mention_everyone: boolean;
  schedule_piket: boolean;
  created_at: Date;
  updated_at: Date;
}

export interface Participant {
  id: ID;
  isMe: boolean;
  name?: string;
  type: Type;
  image?: string;
  isUser: boolean;
  number: string;
  isGroup: boolean;
  pushname: string;
  isBlocked: boolean;
  shortName?: string;
  isBusiness: boolean;
  isMyContact: boolean;
  isWAContact: boolean;
  isEnterprise?: boolean;
  verifiedName?: string;
  verifiedLevel?: number;
  businessProfile?: BusinessProfile;
}
// ...
export interface AdminDetailResponse {
  success: boolean;
  msg: string;
  group: Group;
  status: number;
}
// ...
export interface AdminShufflePfpResponse {
  success: boolean;
  msg: string;
  status: number;
  image: string;
}

export interface CheckIfGroupHasPwResponse {
  success: boolean;
  msg: string;
  value: boolean;
  status: number;
}

export interface CheckIfGroupRegisteredResponse {
  success: boolean;
  msg: string;
  value: boolean;
  status: number;
}

export interface OTPLoginResponse {
  success: boolean;
  msg: string;
  otp: number;
  status: number;
}

export interface FindResponsesTypeResponse {
  success: boolean;
  msg: string;
  responses: Responses;
  status: number;
}

export interface Responses {
  id: string;
  case: string;
  reply: string;
  images: string | null;
  created_at: Date;
  updated_at: Date;
}

export interface APICheckResponse {
  success: boolean;
  msg: string;
  status: number;
}